chore(main): windows rewrite
