--- src/main/windows/about_window.ts
@@ -1,35 +1,0 @@
-import { BrowserWindow, IWindowOptions } from './window';
-import { l10n } from '../l10n';
-import { windows } from './list';
-
-export class AboutWindow extends BrowserWindow {
-  constructor(options: Pick<IWindowOptions, 'onClosed'>) {
-    super({
-      ...options,
-      app: 'about',
-      size: 'small',
-      title: l10n.get('about'),
-    });
-  }
-
-  /**
-   * Create the browser window.
-   */
-  static create() {
-    /**
-     * Don't create if the window exists.
-     */
-    if (windows.about) {
-      windows.about.focus();
-      windows.about.show();
-
-      return;
-    }
-
-    windows.about = new AboutWindow({
-      onClosed: () => {
-        delete windows.about;
-      },
-    });
-  }
-}
