+++ src/main/windows/browser_window.ts
