--- src/main/windows/list.ts
@@ -1,3 +1,0 @@
-import { BrowserWindow } from './window';
-
-export const windows: Assoc<BrowserWindow> = {};
